pyproject.toml
[tool.poetry]
name = "rv-auto-loader"
version = "0.1"
description = "Slingshot RV Plate/v000/LUT/CDL Auto Loader"
authors = ["slingshotvfx <146885925+slingshotvfx@users.noreply.github.com>"]
license = "Apache-2.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "~3.10"
opencolorio = "^2.4.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.5"
pytest = "^8.3.4"
pyside2 = "^5.15.2.1"
types-opencolorio = "^2.2.1.2"

[tool.ruff]
extend-exclude = ["rv"]

[tool.pytest.ini_options]
pythonpath = "src"
addopts = ["--import-mode=importlib"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
